+++ src/lib.rs
@@ -8,7 +8,7 @@ use alloc::vec::Vec;
 use alloc::boxed::Box;
 
 type Position = u8; // huge board (rank 6) have 253 slots, u8 is just perfect.
-type EncodedAction = u64;
+type EncodedAction = u32;
 
 fn encode_action(from: Position, to: Position) -> EncodedAction {
     ((from as EncodedAction) << 8) + (to as EncodedAction)
@@ -26,12 +26,12 @@ pub mod game;
 pub mod mcts;
 
 #[no_mangle]
-pub unsafe extern fn alloc_memory(byte_size: u64) -> *mut u8 {
+pub unsafe extern fn alloc_memory(byte_size: u32) -> *mut u8 {
     vec![0u8; byte_size as _].leak() as *const _ as _
 }
 
 #[no_mangle]
-pub unsafe extern fn free_memory(ptr: *mut u8, byte_size: u64) {
+pub unsafe extern fn free_memory(ptr: *mut u8, byte_size: u32) {
     // Note the second argument is the length. We set it equals to capacity, which may causing droping uninitialized memory if we were dealing with element types that are not u8.
     Vec::from_raw_parts(ptr, byte_size as _, byte_size as _);
 }
@@ -52,18 +52,18 @@ pub unsafe extern fn new_small_game() -> *mut game::Game {
 }
 
 #[no_mangle]
-pub unsafe extern fn get_board_size(game: *mut game::Game) -> u64 {
+pub unsafe extern fn get_board_size(game: *mut game::Game) -> u32 {
     (*game).board_size() as _
 }
 
 #[no_mangle]
-pub unsafe extern fn get_n_pieces(game: *mut game::Game) -> u64 {
+pub unsafe extern fn get_n_pieces(game: *mut game::Game) -> u32 {
     (*game).n_pieces() as _
 }
 
 /// the returned length is only used for deallocation. The `out` length is the same as board size and not explicitly returned.
 #[no_mangle]
-pub unsafe extern fn possible_moves_with_path(game: *mut game::Game, pos: Position, out: *mut *mut Position, length: *mut u64) {
+pub unsafe extern fn possible_moves_with_path(game: *mut game::Game, pos: Position, out: *mut *mut Position, length: *mut u32) {
     let possible_moves_with_path = (*game).possible_moves_with_path(pos);
     *length = possible_moves_with_path.capacity() as _;
     *out = possible_moves_with_path.leak() as *const _ as _;
@@ -71,7 +71,7 @@ pub unsafe extern fn possible_moves_with_path(game: *mut game::Game, pos: Positi
 
 /// returned list is encoded as [INVALID_POSITION, pieces_pos_1, pieces_move_1, pieces_move_2, INVALID_POSITION, pieces_pos_1, ..., INVALID_POSITION, INVALID_POSITION]
 #[no_mangle]
-pub unsafe extern fn all_possible_moves(game: *mut game::Game, out: *mut *mut Position, length: *mut u64) {
+pub unsafe extern fn all_possible_moves(game: *mut game::Game, out: *mut *mut Position, length: *mut u32) {
     let possible_moves = (*game).movable_pieces_and_possible_moves_of_current_player();
 
     let mut encoded = vec![];
@@ -111,7 +111,7 @@ pub unsafe extern fn get_status(game: *mut game::Game) -> u8 {
 /// following 2*`n_pieces` bytes: the position of each pieces, with the first half belongs to the first player.
 /// the returned length is only used for deallocation. The `out` length is determined by 2 + 2 * n_pieces
 #[no_mangle]
-pub unsafe extern fn dump(game: *mut game::Game, out: *mut *mut Position, length: *mut u64) {
+pub unsafe extern fn dump(game: *mut game::Game, out: *mut *mut Position, length: *mut u32) {
     let game = &mut *game;
     let mut encoded = vec![];
 
@@ -142,10 +142,10 @@ pub unsafe extern fn destroy_game(game: *mut game::Game) {
 /// the i-th (in "dump" order) piece of the next player to the position j. The array should be normalized with invalid
 /// actions having probability 0.
 #[no_mangle]
-pub unsafe extern fn new_mcts(policy_cfun: extern fn (*mut game::Game, *mut f64, *mut f64)) -> *mut mcts::Tree {
+pub unsafe extern fn new_mcts(policy_cfun: extern fn (*mut game::Game, *mut f32, *mut f32)) -> *mut mcts::Tree {
     let policy_value_callback = Box::new(move |game: &game::Game| {
         let mut prior = vec![0.0; game.n_pieces() * game.board_size()]; // TODO: uninited?
-        let mut value = f64::NAN; // invalid value to check if `policy_fun` runs correctly
+        let mut value = f32::NAN; // invalid value to check if `policy_fun` runs correctly
 
         policy_cfun(game as *const _ as _, prior.as_mut_ptr(), &mut value as _);
 
@@ -170,12 +170,12 @@ pub unsafe extern fn new_mcts(policy_cfun: extern fn (*mut game::Game, *mut f64,
 }
 
 #[no_mangle]
-pub unsafe extern fn mcts_playout(mcts: *mut mcts::Tree, game: *mut game::Game, ntimes: u64) {
+pub unsafe extern fn mcts_playout(mcts: *mut mcts::Tree, game: *mut game::Game, ntimes: u32) {
     (*mcts).playout(&*game, ntimes as _)
 }
 
 #[no_mangle]
-pub unsafe extern fn mcts_get_action_probs(mcts: *mut mcts::Tree, temp: f64, actions: *mut EncodedAction, probs: *mut f64, length: *mut u64) {
+pub unsafe extern fn mcts_get_action_probs(mcts: *mut mcts::Tree, temp: f32, actions: *mut EncodedAction, probs: *mut f32, length: *mut u32) {
     let action_probs = (*mcts).get_action_probs(temp);
     *length = action_probs.len() as _;
 
@@ -193,7 +193,7 @@ pub unsafe extern fn mcts_get_action_probs(mcts: *mut mcts::Tree, temp: f64, act
 }
 
 #[no_mangle]
-pub unsafe extern fn mcts_sample_action(mcts: *mut mcts::Tree, exploration_prob: f64, temperature: f64) -> EncodedAction {
+pub unsafe extern fn mcts_sample_action(mcts: *mut mcts::Tree, exploration_prob: f32, temperature: f32) -> EncodedAction {
     let (from, to) = (*mcts).sample_action(exploration_prob, temperature);
     encode_action(from, to)
 }
@@ -205,12 +205,12 @@ pub unsafe extern fn mcts_chroot(mcts: *mut mcts::Tree, encoded_action: EncodedA
 }
 
 #[no_mangle]
-pub unsafe extern fn mcts_total_visits(mcts: *mut mcts::Tree) -> u64 {
+pub unsafe extern fn mcts_total_visits(mcts: *mut mcts::Tree) -> u32 {
     (*mcts).total_visits()
 }
 
 #[no_mangle]
-pub unsafe extern fn mcts_root_value(mcts: *mut mcts::Tree) -> f64 {
+pub unsafe extern fn mcts_root_value(mcts: *mut mcts::Tree) -> f32 {
     (*mcts).root_value()
 }
